requirements.txt
# Requirements lower bounds listed here are our best effort to keep them up to
# date but we do not test them so no guarantee of having them all correct. If
# you find any incorrect lower bounds, let us know or propose a fix.
elasticsearch>=8.0.0 # Apache-2.0
falcon>=1.0.0 # Apache-2.0
jsonschema>=3.2.0 # MIT
keystonemiddleware>=4.17.0 # Apache-2.0
Paste>=2.0.2 # MIT
PasteDeploy>=1.5.0 # MIT
oslo.config>=6.8.0 # Apache-2.0
oslo.context>=2.22.0 # Apache-2.0
oslo.db>=4.44.0 # Apache-2.0
oslo.i18n>=3.15.3 # Apache-2.0
oslo.log>=3.36.0 # Apache-2.0
oslo.middleware>=3.31.0 # Apache-2.0
oslo.policy>=4.5.0 # Apache-2.0
oslo.serialization>=2.25.0 # Apache-2.0
oslo.upgradecheck>=1.3.0 # Apache-2.0
SQLAlchemy>=2.0.5 # MIT
alembic>=1.8.0 # MIT
openstacksdk>=1.0.0 # Apache-2.0
